fix: AptInfoResponse 평균 타입 수정

+++ BackEnd/src/main/java/com/ssafy/happy/deal/dto/AptInfoResponse.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class AptInfoResponse {
     private String name;
-    private long average;
+    private double average;
     private String lat;
     private String lng;
 }
